# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервер для интернет-магазина продуктов FoodStore",
        "title": "FoodStore API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/protected/client/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает продукты, отсортированные по приоритету (продавцы с подпиской выше)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Получить продукты по приоритету",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CategoryOutput"
                            }
                        }
                    },
                    "404": {
                        "description": "Продукты не найдены",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/client/address": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет новый адрес доставки клиенту",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Добавить адрес клиента",
                "parameters": [
                    {
                        "description": "Данные адреса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddAddressInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AddAddressOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/client/cart": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает все товары в корзине клиента",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Получить содержимое корзины",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CartItemOutput"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет товар в корзину клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Добавить товар в корзину",
                "parameters": [
                    {
                        "description": "Данные для добавления в корзину",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddToCartInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AddToCartOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/client/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех заказов текущего клиента",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Получить заказы клиента",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ClientOrderOutput"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новый заказ клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Создать заказ",
                "parameters": [
                    {
                        "description": "Данные заказа",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderOutput"
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON или недостаточно товара",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/client/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает полный каталог продуктов по категориям с предложениями продавцов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Получить каталог продуктов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CategoryOutput"
                            }
                        }
                    },
                    "404": {
                        "description": "Продукты не найдены",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/client/products/{product_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает информацию о продукте по его ID с предложениями продавцов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Получить продукт по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID продукта",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductOutput"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/protected/client/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает профиль текущего авторизованного клиента",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Получить профиль клиента",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClientProfileOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddAddressInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "client_id": {
                    "type": "integer"
                }
            }
        },
        "dto.AddAddressOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.AddToCartInput": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "integer"
                }
            }
        },
        "dto.AddToCartOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.CartItemOutput": {
            "type": "object",
            "properties": {
                "cart_id": {
                    "type": "integer"
                },
                "cart_items_id": {
                    "type": "integer"
                },
                "price_at_purchase": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CategoryOutput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sub_category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SubCategoryOutput"
                    }
                }
            }
        },
        "dto.ClientOrderItemDTO": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer"
                },
                "order_items_id": {
                    "type": "integer"
                },
                "price_at_purchase": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "integer"
                }
            }
        },
        "dto.ClientOrderOutput": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ClientOrderItemDTO"
                    }
                },
                "order_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.ClientProfileOutput": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateOrderInput": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateOrderItemInputDTO"
                    }
                }
            }
        },
        "dto.CreateOrderItemInputDTO": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateOrderOutput": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorOutput": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProductOutput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "sellers_offers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SellerOfferOutput"
                    }
                }
            }
        },
...
        "dto.SellerOfferOutput": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "integer"
                },
                "seller_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.SubCategoryOutput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProductOutput"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}